requirements.txt
Django>=5.2,<5.3
djangorestframework>=3.14,<3.15
django-cors-headers>=4.4,<5.0
Pillow>=10.0,<11.0
python-dotenv>=1.0,<2.0
